feat: 비밀번호 입력 페이지 커스텀 훅으로 분리 (#354)
* refactor: 비밀번호 설정 커스텀 훅으로 분리

* refactor: event name 상태 제거

* style: index.ts에 setEventPasswordPage 컴포넌트 추가

* refactor: navigate 책임을 UI 컴포넌트로 위임

* style: 비밀번호 입력을 관리하는 함수라는 의미로 변경
