+++ client/src/pages/CreateEventPage/SetEventPasswordPage.tsx
@@ -2,60 +2,38 @@ import {useEffect, useState} from 'react';
 import {useLocation, useNavigate} from 'react-router-dom';
 import {FixedButton, MainLayout, LabelInput, Title, TopNav, Back} from 'haengdong-design';
 
-import validateEventPassword from '@utils/validate/validateEventPassword';
-import {requestPostNewEvent} from '@apis/request/event';
-
-import useEvent from '@hooks/useEvent';
+import useSetPassword from '@hooks/useSetPassword';
 
 import RULE from '@constants/rule';
 import {ROUTER_URLS} from '@constants/routerUrls';
 
 const SetEventPasswordPage = () => {
-  const [eventName, setEventName] = useState('');
-  const [password, setPassword] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const [canSubmit, setCanSubmit] = useState(false);
-  const {createNewEvent} = useEvent();
   const navigate = useNavigate();
   const location = useLocation();
 
   useEffect(() => {
     if (!location.state) {
       navigate(ROUTER_URLS.main);
-    } else {
-      setEventName(location.state.eventName);
     }
-  }, []);
+  }, [location.state]);
 
-  const submitPassword = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
+  const {password, errorMessage, canSubmit, submitPassword, handlePasswordChange} = useSetPassword(
+    location.state?.eventName,
+  );
 
-    const {eventId} = await createNewEvent({eventName, password: parseInt(password)});
+  const onSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+    const eventId = await submitPassword(event);
 
     navigate(`${ROUTER_URLS.eventCreateComplete}?${new URLSearchParams({eventId})}`);
   };
 
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = event.target.value;
-    const validation = validateEventPassword(newValue);
-
-    setCanSubmit(newValue.length === RULE.maxEventPasswordLength);
-
-    if (validation.isValid) {
-      setPassword(newValue);
-      setErrorMessage('');
-    } else {
-      event.target.value = password;
-      setErrorMessage(validation.errorMessage ?? '');
-    }
-  };
   return (
     <MainLayout>
       <TopNav>
         <Back />
       </TopNav>
       <Title title="행사 비밀번호 설정" description="행사 관리에 필요한 4 자리의 숫자 비밀번호를 입력해 주세요." />
-      <form onSubmit={submitPassword} style={{padding: '0 1rem'}}>
+      <form onSubmit={onSubmit} style={{padding: '0 1rem'}}>
         <LabelInput
           labelText="비밀번호"
           errorText={errorMessage}
@@ -63,10 +41,10 @@ const SetEventPasswordPage = () => {
           type="secret"
           maxLength={RULE.maxEventPasswordLength}
           placeholder="비밀번호"
-          onChange={e => handleChange(e)}
+          onChange={handlePasswordChange}
           isError={!!errorMessage}
           autoFocus
-        ></LabelInput>
+        />
         <FixedButton disabled={!canSubmit}>행동 개시!</FixedButton>
       </form>
     </MainLayout>
