+++ client/src/pages/CreateEventPage/index.ts
@@ -1,2 +1,3 @@
 export {default as SetEventNamePage} from './SetEventNamePage';
+export {default as SetEventPasswordPage} from './SetEventPasswordPage';
 export {default as CompleteCreateEventPage} from './CompleteCreateEventPage';
