+++ client/src/router.tsx
@@ -3,10 +3,9 @@ import {createBrowserRouter} from 'react-router-dom';
 import {AdminPage} from '@pages/EventPage/AdminPage';
 import {HomePage} from '@pages/EventPage/HomePage';
 import ErrorPage from '@pages/ErrorPage/ErrorPage';
-import SetEventPasswordPage from '@pages/CreateEventPage/SetEventPasswordPage';
 import EventLoginPage from '@pages/EventPage/AdminPage/EventLoginPage';
 
-import {CompleteCreateEventPage, SetEventNamePage} from '@pages/CreateEventPage';
+import {CompleteCreateEventPage, SetEventNamePage, SetEventPasswordPage} from '@pages/CreateEventPage';
 import {MainPage} from '@pages/MainPage';
 import {EventPage} from '@pages/EventPage';
